Move DEC helper functions to its own module
